requirements.txt
pandas
matplotlib
seaborn
openai
python-dotenv
pydantic